feat: Refactor DefineGameAreaScreen to use new widget structure for map and overlays
